refactor: create index rollup for Panel Family
