+++ src/Components/Panel/Documentation/PanelExampleA.md
