+++ src/Components/Panel/Documentation/PanelExampleB.md
