+++ src/Components/Panel/Documentation/PanelExampleC.md
