+++ src/Components/Panel/Documentation/PanelFooter.md
