+++ src/index.ts
@@ -72,6 +72,7 @@ export * from './Components/Page/PageHeaderLeft'
 export * from './Components/Page/PageHeaderRight'
 export * from './Components/Page/PageSubTitle'
 export * from './Components/Page/PageTitle'
+export * from './Components/Panel'
 export * from './Components/Panel/Panel'
 export * from './Components/Panel/PanelBody'
 export * from './Components/Panel/PanelFooter'
